refactor: split transformer namespace helpers
